requirements.txt
python-telegram-bot==20.7
pymupdf==1.23.8
pikepdf==8.11.1
Pillow==10.2.0
pdfplumber==0.10.3
openpyxl==3.1.2
img2pdf==0.4.4
reportlab==4.1.0
pytesseract==0.3.10
motor==3.3.2
pymongo==4.6.1
aiohttp==3.9.3
aiofiles==23.2.1
requests==2.31.0
python-dotenv==1.0.0
